# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Radar 2.0 API
  version: 0.1.0
  description: Backend oficial do Radar 2.0 para operacao enterprise em Azure.
servers:
  - url: https://api.radar.example.gov.br
    description: Producao
  - url: https://api-hml.radar.example.gov.br
    description: Homologacao
...
paths:
...
  /v1/teams:
    get:
      tags: [Teams]
      summary: Lista equipe agrupada por microrregiao
      parameters:
        - name: microregionId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Equipe agrupada por microrregiao
    post:
      tags: [Teams]
      summary: Adiciona membro a equipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberRequest"
      responses:
        "201":
          description: Membro criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /v1/teams/status:
    get:
      tags: [Teams]
      summary: Verifica se email ja possui vinculacao com equipe
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Status atual
  /v1/teams/user-municipality:
    post:
      tags: [Teams]
      summary: Salva ou atualiza municipio do usuario na equipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveUserMunicipalityRequest"
      responses:
        "204":
          description: Municipio salvo
  /v1/teams/{memberId}:
    delete:
      tags: [Teams]
      summary: Remove membro da equipe
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Membro removido
...
  /v1/teams/pending-registrations/{id}:
    delete:
      tags: [Teams]
      summary: Exclui pendencia de cadastro
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Pendencia excluida
...
  /v1/actions/{actionUid}/comments:
    get:
      tags: [Comments]
      summary: Lista comentarios de uma acao
      parameters:
        - name: actionUid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lista de comentarios
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
    post:
      tags: [Comments]
      summary: Adiciona comentario em uma acao
      parameters:
        - name: actionUid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Comentario criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionComment"
        "404":
          description: Acao nao encontrada
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/comments/{commentId}:
    patch:
      tags: [Comments]
      summary: Atualiza um comentario
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentRequest"
      responses:
        "200":
          description: Comentario atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionComment"
        "403":
          description: Usuario nao pode alterar o comentario
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Comentario nao encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      tags: [Comments]
      summary: Remove um comentario
      parameters:
        - name: commentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Comentario removido
        "403":
          description: Usuario nao pode remover o comentario
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Comentario nao encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /v1/requests:
    get:
      tags: [Requests]
      summary: Lista requests do usuario, gerenciais ou notificacoes
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [user, managed, notifications]
        - name: limit
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: statusFilter
          in: query
          schema:
            type: string
            enum: [pending, resolved, rejected, all]
        - name: typeFilter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista ou paginacao de requests
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/RequestListResponse"
                  - $ref: "#/components/schemas/ManagedRequestListResponse"
        "401":
          description: Nao autenticado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    post:
      tags: [Requests]
      summary: Cria um novo request de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestRequest"
      responses:
        "201":
          description: Request criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRequest"
        "400":
          description: Payload invalido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
  /v1/requests/{requestId}:
    patch:
      tags: [Requests]
      summary: Atualiza status de um request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestRequest"
      responses:
        "204":
          description: Request atualizado
        "403":
          description: Perfil sem privilegio administrativo
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Request nao encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      tags: [Requests]
      summary: Remove um request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Request removido
        "403":
          description: Perfil sem privilegio administrativo
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Request nao encontrado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  schemas:
...
    TeamMember:
      type: object
      required: [id, microregionId, name, role, email, municipality, isRegistered]
      properties:
        id:
          type: string
        microregionId:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
        municipality:
          type: string
        isRegistered:
          type: boolean
...
    CreateTeamMemberRequest:
      type: object
      required: [microregionId, name, role]
      properties:
        microregionId:
          type: string
        name:
          type: string
        role:
          type: string
        email:
          type: string
          format: email
        municipality:
          type: string
    SaveUserMunicipalityRequest:
      type: object
      required: [microregionId, email, municipality, userName]
      properties:
        microregionId:
          type: string
        email:
          type: string
          format: email
        municipality:
          type: string
        userName:
          type: string
...
    RequestUserSummary:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        cargo:
          type: string
        municipio:
          type: string
        microregiao_id:
          type: string
    UserRequest:
      type: object
      required: [id, user_id, request_type, content, status, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        request_type:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [pending, resolved, rejected]
        admin_notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        resolved_by:
          type: string
          nullable: true
        resolved_at:
          type: string
          nullable: true
        user:
          $ref: "#/components/schemas/RequestUserSummary"
    RequestListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserRequest"
    ManagedRequestListResponse:
      type: object
      required: [items, totalCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/UserRequest"
        totalCount:
          type: integer
    CreateRequestRequest:
      type: object
      required: [requestType, content]
      properties:
        requestType:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [pending, resolved, rejected]
        adminNotes:
          type: string
    UpdateRequestRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, resolved, rejected]
        adminNotes:
          type: string
...
    ActionComment:
      type: object
      required: [id, actionUid, authorId, authorName, authorMunicipio, authorAvatarId, content, createdAt]
      properties:
        id:
          type: string
        actionUid:
          type: string
        parentId:
          type: string
          nullable: true
        authorId:
          type: string
        authorName:
          type: string
        authorMunicipio:
          type: string
        authorAvatarId:
          type: string
        authorRole:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CommentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ActionComment"
    CreateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
        parentId:
          type: string
          nullable: true
    UpdateCommentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string